Move legacy workflow module into src package
