+++ src/core/workflow.py
@@ -1,28 +1,22 @@
-"""
-Main workflow orchestrator for SciResearch Workflow.
-"""
+"""Main workflow orchestrator for SciResearch Workflow."""
 from __future__ import annotations
-import sys
+
 import logging
-from typing import Optional
-from pathlib import Path
-
-# Import modular components
-from core.config import WorkflowConfig, DEFAULT_MODEL
-from core.quality import QualityAssessment
-from ai.chat import AIChat
-from ai.prompts import PromptTemplates
-from processing.latex import LaTeXProcessor
-from processing.files import FileManager
-
-# Import existing workflow functions from legacy workflow
 import sys
 from pathlib import Path
-# Add the parent directory to the path for imports
-sys.path.insert(0, str(Path(__file__).parent.parent.parent))
+from typing import Optional
 
-# Import functions from legacy workflow
-from legacy_monolithic_workflow import _initial_draft_prompt, run_simulation_step, run_optimized_review_revision_step
+from src.ai.chat import AIChat
+from src.ai.prompts import PromptTemplates
+from src.core.config import DEFAULT_MODEL, WorkflowConfig
+from src.core.quality import QualityAssessment
+from src.legacy.legacy_monolithic_workflow import (
+    _initial_draft_prompt,
+    run_simulation_step,
+    run_optimized_review_revision_step,
+)
+from src.processing.files import FileManager
+from src.processing.latex import LaTeXProcessor
 
 logger = logging.getLogger(__name__)
