+++ src/legacy/__init__.py
@@ -1,0 +1,3 @@
+"""Legacy workflow implementations for SciResearch."""
+
+__all__ = ["legacy_monolithic_workflow"]
